refactor: rename storage package to persistence
