+++ app/src/main/java/com/monstarlab/features/resources/domain/repository/ResourceRepository.kt
@@ -2,7 +2,7 @@ package com.monstarlab.features.resources.domain.repository
 
 import com.monstarlab.features.resources.data.api.ResourcesApi
 import com.monstarlab.features.resources.data.api.dtos.toResource
-import com.monstarlab.features.resources.data.storage.ResourcePreferenceStore
+import com.monstarlab.features.resources.data.persistence.ResourcePreferenceStore
 import com.monstarlab.features.resources.domain.model.Resource
 import javax.inject.Inject
