+++ app/src/main/java/com/monstarlab/features/user/data/repository/UserRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.monstarlab.features.user.data.repository
 
 import com.monstarlab.features.user.data.api.UsersApi
 import com.monstarlab.features.user.data.api.dtos.toUser
-import com.monstarlab.features.user.data.storage.UserPreferenceStore
+import com.monstarlab.features.user.data.persistence.UserPreferenceStore
 import com.monstarlab.features.user.domain.model.User
 import com.monstarlab.features.user.domain.repository.UserRepository
 import javax.inject.Inject
